refactor(internal/rust): merge rust packages
Merge the internal/librarian/rust package into
internal/librarian/internal/rust to consolidate rust functionality into
a single rust package.

Remove the typo checks from internal/sidekick to maintain feature
parity.
