--- internal/librarian/internal/rust/testdata/Cargo.toml
@@ -1,26 +1,0 @@
-[workspace]
-members = ["google-cloud-secretmanager-v1"]
-resolver = "2"
-
-[workspace.package]
-edition = "2024"
-authors = ["Google LLC"]
-license = "Apache-2.0"
-repository = "https://github.com/googleapis/google-cloud-rust/tree/main"
-keywords = ["gcp", "google-cloud", "google-cloud-rust", "sdk"]
-categories = ["network-programming"]
-rust-version = "1.85.0"
-
-[workspace.dependencies]
-async-trait = "0.1"
-bytes = "1"
-gax = { version = "1", package = "google-cloud-gax" }
-iam_v1 = { version = "1", package = "google-cloud-iam-v1" }
-location = { version = "1", package = "google-cloud-location" }
-reqwest = "0.12"
-serde = "1"
-serde_json = "1"
-serde_with = "3"
-time = "0.3"
-tokio-test = "0.4"
-wkt = { version = "1", package = "google-cloud-wkt" }
