+++ internal/librarian/rust/generate_test.go
