+++ internal/librarian/rust/sidekick_config_mapper.go
